Move each class/enum to its own file
