+++ src/EDXLNet/Models/DistributionStatus.cs
@@ -1,30 +1,7 @@
-﻿using System;
-
-namespace EDXLNet.Models
+﻿namespace EDXLNet.Models
 {
     public class DistributionStatus
     {
         public StatusKindDefault StatusKindDefault { get; set; }
-    }
-
-    public class StatusKindDefault
-    {
-        public Uri ValueListUri
-        {
-            get
-            {
-                return new Uri("urn:oasis:names:tc:emergency:EDXL:DE:2.0:Defaults:StatusKind");
-            }
-        }
-
-        public StatusKindDefaultValues Value { get; set; }
-    }
-
-    public enum StatusKindDefaultValues
-    {
-        Actual,
-		Exercise,
-		System,
-		Test
     }
 }
