+++ src/EDXLNet/Models/StatusKindDefaultValues.cs
@@ -1,0 +1,10 @@
+﻿namespace EDXLNet.Models
+{
+    public enum StatusKindDefaultValues
+    {
+        Actual,
+        Exercise,
+        System,
+        Test
+    }
+}
